I cleaned up some code Made saving better(why did I do it like that?) yuh

+++ main.py
@@ -1,43 +1,67 @@
 import sys
 import os
-if os.path.exists("requirements.txt"):
+
+CONFIGFILE = "config.toml"
+if os.path.exists("requirements.txt") and not os.path.exists(CONFIGFILE):
 	os.system("pip install -r requirements.txt")
-	os.remove("requirements.txt")
 
 import tkinter as tk
 import requests
 import utils
+import tomlkit as toml
 
 try:
 	os.mkdir("walltaker")
 except FileExistsError:
 	pass
 
-try:
-	requests.get("https://walltaker.joi.how/")
-except ConnectionError as e:
+def disconnected(e, quit=False):
 	win = tk.Tk()
 	win.title("Error")
 	label = tk.Label(win, text="Connection error")
 	statusL = tk.Label(win, text=e)
+	label.pack()
+	statusL.pack()
 	win.mainloop()
-	quit(0)
+	if quit:
+		quit(0)
 
 win = tk.Tk()
 win.title("Walltaker YClient")
 
 interval	= 60
-linkID		= 14532
-if os.path.exists("sav"):
-	with open("sav", "rb") as file:
-		interval = int.from_bytes(file.read(4), sys.byteorder)
-		linkID = int.from_bytes(file.read(4), sys.byteorder)
-prevcall = 0
-def new_wallpaper():
-	global prevcall, currentwallpaper
+linkID		= 0
 
-	link = requests.get(f"https://walltaker.joi.how/links/{linkID}.json").json()
-	post = link["post_url"]
+if os.path.exists(CONFIGFILE):
+	with open(CONFIGFILE, "r") as file:
+		save = toml.load(file)
+		interval = save["client"]["interval"]
+		linkID = save["user"]["linkID"]
+
+def new_wallpaper():
+	global prevcall
+	win.after_cancel(prevcall)
+	prevcall = win.after(interval*1000, new_wallpaper)
+	try:
+		link = requests.get(f"https://walltaker.joi.how/links/{linkID}.json",
+						headers={"User-Agent": "Yuri-client"},
+						timeout=10)
+	except requests.exceptions.ConnectionError as e:
+		disconnected(e)
+		win.after_cancel(prevcall)
+		prevcall = win.after(10_000, new_wallpaper)
+		return
+		
+	link = link.json()
+	
+	try:
+		post = link["post_url"]
+	except KeyError:
+		disconnected("Invalid linkID")
+		win.after_cancel(prevcall)
+		prevcall = win.after(10_000, new_wallpaper)
+		#return
+
 	filename = post.split("/")[-1]
 
 	if os.path.exists(f"walltaker/{filename}"):
@@ -50,7 +74,7 @@ def new_wallpaper():
 	print(utils.getBackground())
 
 def update(newInterval, newLinkID):
-	global interval
+	global interval, linkID
 	try:
 		interval = int(newInterval)
 		linkID = int(newLinkID)
@@ -80,13 +104,19 @@ linkIDI.grid(row=0, column=1)
 updateB.grid(row=1, column=0)
 quitB.grid(row=1, column=1)
 
-win.after(1_000, new_wallpaper)
+prevcall = win.after(1_000, new_wallpaper)
 win.mainloop()
 
-with(open("sav","wb")) as file:
-	interval = bytes(int.to_bytes(interval, 4, sys.byteorder))
-	linkID = bytes(int.to_bytes(linkID, 4, sys.byteorder))
-	file.write(interval+linkID)
+with(open(CONFIGFILE,"w")) as file:
+	save = {
+		"client": {
+			"interval": interval
+		},
+		"user": {
+			"linkID": linkID
+		}
+	}
+	toml.dump(save, file)
 
 print("Exiting")
 exit(0)
